requirements.txt
# Core agent dependencies
langgraph>=0.2.0
langchain-openai>=0.2.0
langchain-core>=0.2.0
pydantic>=2.6.0
python-dotenv>=1.0.0

# Data processing (for build_sanitized_data.py)
pandas>=2.0.0
